Remove some uses of snprintf
Rather than use snprintf to convert the error code, use std::ostringstream since
Windows does not have snprintf.  Use std::to_string for the conversion of the
user id to a string value.  Since on Windows a uid is not an integer, provide an
overload for std::to_string for the PSID.