build,documentation: remove Windows Phone/Store
The Windows Phone platform is no longer being shipped and the Windows Store
platform is no longer in the same state as it was when this project was under
development.  The architectural support is important to retain as Windows on ARM
and Windows ARM64 are still available and usable, though the store platform is
closer to UWP and WinRT now.  The difference means that it may be possible to
use a more traditional model with multiple processes.  This simplifies the build
matrix can reduce the build complexity.