Host: repair the build with `-Werror=stringop-truncation`
When building with GCC 9.3, the `strncpy` is flagged as a possible truncation as
we use the size of the destination for the `strncpy` which expects space for the
trailing nul.  However, we are working with a fixed size buffer, and so prefer
the use of `memcpy` instead.

Take the opportunity to explicitly write the leading nul-character for the unix
path.