documentation: further simplify the README
The build instructions for all the platforms is roughly equivalent at this
point.  The difference now is in the dependencies where Windows recommends
WinFlexBison over flex and bison to avoid MinGW for native builds.
Additionally, the native Windows build requires Visual Studio.  However, the
build steps are identical as we can now rely on CMake to generate the Visual
Studio/msbuild projects.