Host: remove unnecessary Windows API checks and shims
These functions are not currently used in the implementation.  Remove them to
reduce the configure time.  Additionally, since the minimum version of Windows
which makes sense to support currently is Windows 10 (with the older Windows now
being EOL'ed), we can likely reduce the checks and shims further.  This removes
the easiest cases first.