RISCV: improve readablity of RISCV disassembler
Replace the bit-extraction operations with a helper template type `bits` which
takes two template parameters to indicate the bits to extract.  This improves
the readability and maintainability of the RISCV disassembler.  The idea itself
was something that was suggested by @dmlockhart in #47.  This noticably improves
the ability to cross-reference the RISCV ISA specification.