[android] list of lldb tests that fail on ds2 Android-x86_64
This is the complete list of lldb tests that we need to skip when
running the lldb tests against ds2 on Android-x86_64. It is a superset
of some of the other exclusion lists and should eventually be pruned
down and partially refactored into some of the other arch- and
platform-specific exclusion files.

Note: all tests liste here are "skip" rather than "xfail" because many
of them hang during execution and some are flaky and only occasionally
fail. Skipping is the only way to get reliabl tests until we fix test
and/or ds2 issues.