build: reduce altering `CMAKE_{C,CXX}_FLAGS`
These variables are reserved by CMake to be user-controlled. We were
previously modifying these to inject flags for the sanitizer and
coverage builds. Use `add_compile_options` to control the default global
options for all subdirectories. This allows the user to continue to
properly control the C/C++ compile options.