build: simplify and unify macros for Windows
When building for Windows, we define the minimum Windows version that we
run on as Windows Vista, exclude the `MIN` and `MAX` macros as they
conflict with `std::min` and `std::max`, and minimise the windows
type definitions with `WIN32_LEAN_AND_MEAN`. Add the definitions in a
single invocation.