Session: properly handle vFile:open: flags
The flag mapping for `vFile:open:` commands implemented in
`Session::ConvertOpenFlags` was completely wrong for LLDB mode. The
actual flag definitions are found
[here](https://github.com/llvm/llvm-project/blob/main/lldb/include/lldb/Host/File.h).
I don't think this ever worked unless the values were changed in lldb
since this was implemented.

The miss-mapping caused a large number of the lldb tests to fail when
run against ds2 on Linux because it couldn't create + write new files.