Support: clear out some of the scripts
Remove the Docker image creation rules (Dockerfile) for the CircleCI
docker image. The image has been unmaintained for long enough that it is
not valuable any longer. Furthermore, with the migration to GitHub
actions, the image is no longer something we need to construct as GitHub
provides the image.

The helper scripts that were only in use for the Docker image can also
be removed. The only script here that is potentially of particular
interest is the `prepare-android-ndk.py` which is, in effect, a
specialised invocation of `curl`.