[bugfix] propertly handle vCont all threads identifier -1
## Issue
The ds2 `vCont` implementation does not correctly handle actions with
thread-id `p<pid>.<tid>` where `<tid>` is `-1`. A value of `-1`
indicates the action should be applied to *all* threads, but ds2
processes the request on only the *current* thread. This bug causes many
of the multi-threaded lldb tests to hang when run against ds2.

Documentation for `vCont` is
[here](https://sourceware.org/gdb/current/onlinedocs/gdb.html/Packets.html#vCont-packet),
and details on thread-id syntax are
[here](https://sourceware.org/gdb/current/onlinedocs/gdb.html/Packets.html#thread_002did-syntax).

## Fix
Fixing this issue involves several changes:
1. Cleanup `ProcessThreadId::parse` so it will properly fail when
`stroul` fails to parse an integer value.
* It was relying on `errno` to detect `stroul` failure, but `errno` is
only guaranteed set on over/underflow errors so this check is
insufficient. Now it also compares the output and input string positions
to see if any chars were actually consumed.
* Delete the sketchy `CHECK_AND_RESET` macro and only update the member
`pid` and `tid` values on success so the macro isn't needed.
2. Detect the `p<pid>.<tid>` thread-id format when handling `vCont`
packets and force the pid/tid parsing to use GDB multi-process mode.
Without this change, `ProcessThreadId::parse` fails when the server is
operating in LLDB mode due to the fixes in # 1. Does not change the
behavior of `ProcessThreadId::parse` because it is used for commands
other than `vCont` that have alternate syntax.
3. When an action in a `vCont` request has `tid` of `-1`, treat it as a
global action and use the existing codepath to process it. Implement
additional error checking to ensure `pid` matches the debugged process
if valid.

## Test Plan

Run the multi-threaded lldb tests (found under
llvm-project/lldb/test/API/functionalities/thread) against ds2 on Linux
and Android. The tests no longer hang (there are still some failures).

Adds a few more tests to the exclusion list which are now flaky. A
subsequent PR will re-enable many more tests.