[android] use procfs to check pid validity
The POSIX implementation of `Process::isAlive()` relies on succesfully
calling `kill(pid, 0)` to validate the targed process' PID. This call
fails with `EPERM` when ds2 is running in an Android sandbox on newer
versions of Android.

Fix the issue by instead checking that `/proc/PID` exists. Perform this
check only in the Linux override of `Process::isAlive()` since using
procfs is not portable. There is no reason to scope the change to be
Android-only, though, since it will work on any Linux system.

It is worth noting that any solution to validate a PID is best-effort
and is inherently racy. We could technically eliminate the check
entirely and ds2 will still work-- it just won't fail as fast when
trying to attach to an invalid PID.