[freebsd] fix the freebsd build (#116)
I wanted to verify some Linux-specific changes don't break other
platforms and found that ds2 doesn't currently build on FreeBSD 14.0.
These patches address the issues encountered.

The most noteworthy change is the addition of `file(MAKE_DIRECTORY)`
calls to the `RegsGen2` steps. This step was failing on FreeBSD because
the directories did not already exist. Not sure why it succeeds on other
platforms, but the `MAKE_DIRECTORY` addition should be harmless when not
needed.

### Validation
Built on VM running 14.0-RELEASE-p10.
Launched ds2 on FreeBSD in platform mode, verified lldb can connect.

Co-authored-by: Saleem Abdulrasool <compnerd@compnerd.org>