Session support for `qModuleInfo` platform command
Add support for the `qModuleInfo` platform command packet to the ds2
platform session implementation. The generic session handler for this
command was already implemented, but there was no platform or debug
session implementation so it never succeeded.

Adds a new `ModuleInfo` type to hold the information required in the
response packet to `qModuleInfo`. The previous implementation
incorrectly assumed the existing `SharedLibraryInfo` type would suit
this purpose but it does not.