[linux] step kReasonNone events in Linux::Process::wait
The scenario where a `kEventNone` event with a reason other than
`kReasonThreadSpawn` is alluded to in the comment block of
`Linux::Process::wait` but it suggests it will never happen while
stepping. However, this scenario can legitimately occur when a
thread is stepping and another thread in the process hits a
breakpoint.

This change reworks the `Linux::Process::wait` function to be more
readable and fixes the case where a thread stops with `kEventNone`
and `kReasonNone` while single stepping.