[linux] handle stop events while stepping
## Purpose
Properly handle a case that occasionally triggers
[this](https://github.com/compnerd/ds2/blob/main/Sources/Target/Linux/Process.cpp#L274)
`DS2BUG` fatal error in `Linux::Process::wait` when running some
multi-threaded lldb tests.

## Overview
* Handle the case where we get a stop event for a single-stepping thread
with `kEventNone` and `kReasonNone`
* Enable 15 lldb test cases that were previously flaky due to this bug
* Rework the `Linux::Process::wait` function be a little bit more
readable
* Handle all failure cases in `Linux::Process::wait` with the `CHK()`
macro

## Problem Details
The scenario where a `kEventNone` event with a reason other than
`kReasonThreadSpawn` is alluded to in
[this](https://github.com/compnerd/ds2/blob/main/Sources/Target/Linux/Process.cpp#L236-L238)
comment block of `Linux::Process::wait` but it suggests it will never
happen while the thread is single stepping. However, this scenario can
legitimately occur and is fairly easy to reproduce with the lldb
multithreaded tests. In this case, the reason will be `kReasonNone`.

The implementation already handled this `kEventNone` and `kReasonNone`
event by continuing the thread as long as it wasn't also being
single-stepped. This change updates it to step the thread in the same
situation if it is being single-stepped. However, I haven't 100%
convinced myself the implementation was correct to begin with (it could
probably use an overhaul). Intuitively, I would expect to ignore events
with `kEventNone` and `kReasonNone`, but that's not what it was doing
previously.

## Validation
The re-enabled tests now pass consistently locally on Android and Linux,
and have now passed on at least 10 automated runs.

Manually set some breakpoints and stepped in a simple multi-threaded
program. It behaved as expected.

Co-authored-by: Saleem Abdulrasool <compnerd@compnerd.org>