[linux] handle Ctrl-C interrupt packet when inferior is already stopped
## Purpose
Properly support interruption from the debugger via Ctrl-C (\x03) packet
when no thread in the inferior process is running

## Overview
* Implement a `sendInterrupt` method in the `POSIX::Process` class that
can be used to fork a short-lived process whenever an interrupt packet
is received. The forked process immediately exits, generating a process
exit event.
* Implement a `checkInterrupt` method in the `POSIX::Process` class that
compares a provided pid against the pid that was memoized on the last
`sendInterrupt` call. If they match, it clears the the pid and returns
true.
* Wire-up `sendInterrupt` to the `Linux::Process:interrupt` method.
There was previously no Linux override for this `Process` method, so
there's a bit of additional scaffolding for this.
* Wire-up `checkInterrupt` to the `Linux::Process::wait` method and
properly handle the interrupt.
* Enable two lldb tests on Android that were previously failing due to
this bug.

NOTE: This PR only addresses Linux and Android, but the bug most likely
impacts Darwin and FreeBSD platforms as well. The implementations of
`sendInterrupt` and `checkInterrupt` are in `POSIX::Process` so they can
be reused on these platforms.

## Problem Details
The ds2 main thread blocks indefinitely on a `::waitpid()` call from
`Linux::Process::wait()`. In this scenario, if the inferior process
never generates any process events, ds2 is effectively hung. However,
the debug server needs to always be able to be interrupted with a Ctrl-C
(\x03) packet from the debugger. Ds2 already properly handles this
command (and only this command) off of the main thread attempting to
stop the inferior process with a `kill()` call. However, the `kill()`
call is a noop if all threads in the inferior process are already
stopped. In this scenario the lldb <-> ds2 connection is effectively
hung.

## Validation
The re-enabled tests now pass consistently locally on Android and Linux.