support the qModuleInfo packet in debug sessions
## Purpose
Properly support the `qModuleInfo` packet in both platform and debug
sessions.

## Overview
* Move the `onModuleInfo` function from `PlatformSessionImpl` into
`FileOperationsMixin`, making it available in both `PlatformSessionImpl`
and `DebugSessionImpl`
* Re-enable lldb tests that are fixed by this change

## Problem Details
A lot of lldb tests rely on unwinding through stack frames in standard
libraries like `libc.so` and `libc++.so`. It turns out ds2 was not
properly resolving symbols in these libraries because it failed to query
their module info and therefore lldb never copied them into module
cache. This issue was addressed for processes running in Android
sandboxes when `qModuleInfo` support was added in
246c9c01e2946a4192c75b8fb0a0cfe9e477792e. However, this implementation
was only wired-up in the **platform** session implementation and not in
**debug** session. Apparently lldb uses `qModuleInfo` in both session
contexts. By moving the implementation of `onQueryModuleInfo` into the
shared `FileOperationsMixin` class, it becomes available in both session
types.

## Validation
Locally verified the re-enabled tests pass on an Android x86_64 emulator