explicitly map host name of * to 0.0.0.0 for compatibility
On Fedora 40, `getaddrinfo()` resolves host name * to 0.0.0.0. However,
on Android it does not, which leads to a number of lldb tests failling
when run against a ds2 server. This change makes explicitly remaps the
name prior to calling `getaddrinfo()` to ensure the most compatibility.

This change matches the behavior of lldb-server, which was added (back)
in llvm/llvm-project@f57453aece3e1dc56ce00a29ec6612d33e521c88 to address
similar test failures.