add support for the vFile:fstat packet
## Purpose
Add support for lldb's `vFile:fstat` packet to ds2 debug and platform
sessions. Fixes #135.

## Overview
* Add `fstat` method to the `File` class and implement it for POSIX.
Leave the Windows implementation unimplemented (as is the rest of
`File`)
* Call the new `File::fstat` method when is `vFile:fstat` packet is
received.

## Problem Details
Several of the lldb tests rely on `vFile:fstat` to fetch files from the
device, and these tests fail against ds2 because it is not implemented.

The `vFile:fstat` packet isn't documented. The implementation is based
on the expectation of the lldb test case
`TestGdbRemotePlatformFile.TestGdbRemotePlatformFile.test_platform_file_fstat_llgs`,
which now passes with this change.

## Validation
Once this PR and a some additionalchanges are merged, most of the
`TestGdbRemotePlatformFile` tests succeed with ds2 running on Android
and Linux and can be enabled in CI. They do not all pass with this PR
alone.

Co-authored-by: Saleem Abdulrasool <compnerd@compnerd.org>