fix the platform session process launch implementation (#143)
## Purpose
Add proper support for the `A` packet (program arguments) in the
platform session implementation, launching programs on request from the
connected debugger. Along with PR #146, addresses issue #137.

## Overview
* Implement `ProcessLaunchMixin::onSetProgramArguments` to spawn the
specified executable with stdio and environment variables memoized in
previous calls to `onSetStdFile` and `onSetEnvironmentVariable`. Use the
existing `ProcessSpawner` implementation for the heavy lifting.
* Implement `ProcessLaunchMixin::onQueryProcessInfo` to enable the
debugger to get details about the most recently launched process. Reuse
the exiting `Platform::GetProcessInfo` function to retrieve the process
information.
* Rename the pid-based `SessionDelegate::onQueryProcessInfo` method to
`SessionDelegate::onQueryProcessInfoPID` to fix a `-Woverloaded-virtual`
compiler warning. This also makes the method name consistent with its
associated command packet `qProcessInfoPID` (as `onQueryProcessInfo` is
with the `qProcessInfo` packet).
* Implement `ProcessLaunchMixin::onTerminate` to kill the specified
process and wait for it to exit. This method is invoked when a `vKill`
packet is received. Only allow processes that were previously launched
via this mechanism to be killed in this way by maintaining a list of all
launched processes.
* Implement the POSIX version of a new `Platform::TerminateProcess`
function, which uses `kill(2)` and `waitpid(2)` to kill the requested
process.
* Stub-out the Windows version of `Platform::TerminateProcess` for
future implementation. There is no platform mode for Windows, so this
code path is not currently reachable.
* Special-case the scenario where the program being launched is ds2 in
gdbserver mode and propagate logging settings to make it easier to debug
test failures.


## Problem Details
Typically, the `qLaunchGDBServer` packet is used to launch another
instance of ds2 running in "gdb server" mode. Many of the lldb tests use
this mechanism. However, some tests use the `A` (program arguments)
packet to launch new ds2 instances instead, and this method doesn't work
correctly when ds2 is running in "platform" mode.

Currently, ds2 uses the `ProcessLaunchMixin` implementation in
`PlatformSession` to ostensibly support launching processes. However, it
only memoizes the program details (environment, stdio files) and never
actually launches the program even though it reports success to the
connected debugger.

It is worth noting that the `A` packet *is* handled correctly by
`DebugSession`, but that implementation is specific to launching
inferiors to be debugged (setting up ptrace, etc). `PlatformSession`
cannot share the same implementation because platform mode launches
programs without the intent of debugging them. They can both use
`ProcessSpawner` for the bulk of the implementation.

## Validation
Once this PR and #146 are merged, most of the `TestGdbRemote*` and
`TestLldbGdbServer*` lldb tests succeed with ds2 running on Android and
Linux and can be enabled in CI. They do not all pass with this PR alone,
but local validation confirms the following:
* Child debug server instances are launching in response to `A` packets
and connect to the debugger/test over independent sockets
* Child processes are being killed and cleaned up in response to `vKill`
packets received during test cleanup