implement a subset of match filters for the `qfProcessInfo` and `qsProcessInfo` packets (#168)
## Purpose
Properly implement a subset of match filters for the `qfProcessInfo` and
`qsProcessInfo` packets.

fixes #167
fixes #171 

## Overview
* Fix the `Handle_qfProcessInfo` implementation to parse the incoming
string; it was mistakenly using the full argument string `args` instead
of the individual argument string `arg`.
* Decode the hex-encoded process name string to match against in
`Handle_qfProcessInfo`. It was incorrectly not decoded before.
* Fix `PlatformSessionImplBase::updateProcesses` to clear the process
list before enumerating/matching.
* Without this fix, the process list keeps growing with redundant
entries.
* Implement `PlatformSessionImplBase::processMatch` to perform the
actual matching against the values provided with the `qfProcessInfo`
packet. They were previously ignored (and noted in a TODO).
* When matching against the process name, try matching against both the
process name in the `ProcessInfo` structure and the base thread name for
the process.
* This matching is necessary on Linux and Android because the process
name is the fully qualified path name to the executable, but the match
name is just the base name (e.g. "/home/user/a.out" vs "a.out") which
matches the primary thread name.
* This change was the most targeted fix that didn't disrupt other
behavior and break other test cases that rely on `ProcessInfo.name`
being the fully qualified executable file name.
* Add TODOs for the remaining missing functionality. The test coverage
on this code path appears to be incomplete.
* Enable the
`TestProcessAttach.ProcessAttachTestCase.test_attach_to_process_by_name`
test case on Linux and Android.

## Background
The `qfProcessInfo` and `qsProcessInfo` packets are and LLDB platform
extension and are used together to enumerate processes on the target.
They are documented
[here](https://lldb.llvm.org/resources/lldbgdbremote.html#qfprocessinfo-qsprocessinfo-platform-extension).

## Validation
Ranually run the `test_attach_to_process_by_name` test case on Android
and Linux to confirm it passes.